+++ docs/guide/changelog.md
@@ -3,6 +3,11 @@
 <BackTop />
 <Watermark fullscreen content="Vue Amazing UI" />
 
+## 1.4.2 <Tag color="volcano" size="small">2024-08-01</Tag>
+
+- 优化并更新 [滚动条 Scrollbar](https://themusecatcher.github.io/vue-amazing-ui/guide/components/scrollbar.html) 组件，新增 `contentClass` 属性
+- 优化 [选择器 Select](https://themusecatcher.github.io/vue-amazing-ui/guide/components/select.html) 组件
+
 ## 1.4.1 <Tag color="volcano" size="small">2024-07-31</Tag>
 
 - 优化 [按钮 Button](https://themusecatcher.github.io/vue-amazing-ui/guide/components/button.html) 组件